typo in initial user get route....quick fix
